requirements.txt
picamera
flask
Pillow
